requirements.txt
bounding_box==0.1.3
matplotlib==3.10.7
numpy==2.3.5
onnxruntime==1.23.2
onnxruntime_extensions==0.14.0
onnxruntime_gpu==1.23.2
opencv_contrib_python==4.12.0.88
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
Pillow==12.0.0
torch==2.6.0+cu124
torchvision==0.21.0+cu124
